requirements.txt
flask==2.3.3
pytesseract==0.3.10
pillow==10.0.0
psycopg2-binary==2.9.6
gunicorn==21.2.0